Update PSTTable7C.java
Edited Line 209.

The logical operator is changed to greater than symbol.
currentValueArrayStart + TCI_bm > rowNodeInfo.length()

This is in connection to Issue #12, where we get ArrayIndexOutOfBound exception because few recipients will skipped because of the wrong logical operator.
Result of which is itemList will have size less than the actual no of recipients. And when we try to fetch we get arrayIndexOutOFBound.

ArrayStart if less than the actual row length then it should proceed forward and added to the itemsList. But previously few are getting skipped(because of continue). 

This patch will fix the issue.

Thank you. 

Regards,
Ajay Sadhu