Handle case where there's no data for an attachment
This changes the getFileInputStream method to return an empty InputStream when there is no data for an attachment. Returning an empty stream instead of null allows downstream parsers like Tika to avoid having to check for null for explicitly and properly or semantically reflects the fact that there is a stream, but it contains no data.